p4/ui: tear-free double-buffered DSI + anti-aliased, filled flow graph
Display (P4 DSI path only; other boards untouched):
- Replace Arduino_GFX's single-framebuffer DSI wrapper with direct esp_lcd:
  DPI panel owns two PSRAM framebuffers. LVGL stays in PARTIAL mode; the
  flush rotates strips into the back buffer (same mapping as Arduino_GFX
  rotation=1, so orientation/touch are unchanged), presents via zero-copy
  scan-out flip (latched by the driver at the frame boundary), and defers
  the flip-wait + dirty-row front->back sync to the start of the next frame
  where the wait is effectively free. Kills the flush-during-scan blink the
  flow graph had; deliberately avoids LVGL direct-mode sync (see the S3 4.3B
  post-mortem comment for why that path was rejected).
- Measured on hardware (rolling graph): 26-30 fps, CPU 18%->29%, flush
  4|10 ms vs the blocking variant's 4|18 ms.

Flow graph (all boards):
- Anti-aliased trace: float pen position, coverage-feathered vertical
  segments (~1.5px soft line) instead of a hard 1px run that visibly steps.
- Translucent fill (~40% accent) under the line, gridlines blended through.

+++ include/ui/home_tab.h
@@ -70,7 +70,7 @@ struct HomeWidgets {
   int flow_w = 0;             // plot width/height in px (canvas size)
   int flow_h = 0;
   int flow_stride = 0;        // buffer row pitch in pixels (>= flow_w)
-  int flow_prev_y = -1;      // previous sample's row (continuous line; -1 = pen up)
+  float flow_prev_y = -1.0f;  // previous sample's row, fractional (AA); -1 = pen up
   uint32_t flow_tick = 0;    // last advance timestamp (ms); 0 = uninitialized
   uint32_t flow_accum_ms = 0;  // fractional-pixel time bank
   bool flow_blanked = false;   // true while the plot is cleared (scale disconnected)
